+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1-3] - 2020-10-16
+## \[1.1-4\] - 2021-01-29 11:04
+
+### Added
+
+- debug option
+- log option
+- logging description to the readme file
+
+## \[1.1-3\] - 2020-10-16
 
 ### Added